requirements.txt
# Core Dependencies
python-dotenv==1.0.1
pydantic==2.11.9
pydantic-settings==2.7.0

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.35.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.36
alembic==1.13.0

# Browser Automation
playwright==1.55.0
selenium==4.27.0
browser-use==0.7.8
stagehand==0.5.3

# Web Scraping
beautifulsoup4==4.12.3
httpx==0.28.1
httpx-sse==0.4.1
requests==2.32.0
lxml==5.3.0
firecrawl-py>=1.0.0

# AI/LLM Integration
openai==1.108.0
groq==0.31.1
langchain==0.3.17
langchain-openai==0.3.33

# Google Sheets Integration
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.3.0

# Data Processing
pandas==2.2.3
numpy==2.2.2
openpyxl==3.1.5
PyPDF2==3.0.1
python-dateutil==2.9.0

# Task Queue & Scheduling
celery==5.3.4
redis==5.0.1
apscheduler==3.11.0
schedule==1.2.2

# Development Tools (Optional - for contributors only)
# Uncomment if you want to contribute to the project
# pytest==8.4.1
# pytest-asyncio==0.24.0
# pytest-cov==6.0.0
# black==25.1.0
# flake8==7.2.0
# mypy==1.14.1

# Utilities
python-json-logger==3.2.1
tenacity==9.0.0
click>=8.1.0